requirements.txt
asgiref==3.8.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.3
Django==4.2.20
django-ckeditor==6.7.3
django-cors-headers==4.4.0
django-extensions==3.2.3
django-js-asset==3.1.2
django-ratelimit==4.1.0
django-recaptcha==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
graphviz==0.20.3
idna==3.10
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydot==3.0.4
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
whitenoise==6.9.0
xlrd==2.0.1

# Additional dependencies for Gawulo offline-first system
django-cacheops==7.2
django-redis==5.4.0
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django-filter==23.5
django-guardian==2.4.0
django-storages==1.14.2
boto3==1.34.0
qrcode==7.4.2
reportlab==4.0.7

# Database URL parsing
dj-database-url==2.1.0

# Development dependencies
gunicorn==21.2.0
